csr-git-ops-pipeline-status/requirements.txt
altgraph==0.17
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
gitdb==4.0.5
GitPython==3.1.8
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-build==2.0.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
libcst==0.3.10
macholib==1.14
mypy-extensions==0.4.3
proto-plus==1.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
smmap==3.0.4
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
